Implemented profile plotting
